+++ backend/src/feeding-tasks/feeding-tasks.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { FeedingTasksController } from './feeding-tasks.controller';
+import { FeedingTasksRepository } from './feeding-tasks.repository';
+import { FeedingTasksService } from './feeding-tasks.service';
+
+@Module({
+  controllers: [FeedingTasksController],
+  providers: [FeedingTasksRepository, FeedingTasksService],
+})
+export class FeedingTasksModule {}